changed all double declaration for N's to ints and cast double on it in h's

+++ numint.c
@@ -106,9 +106,9 @@ double adapt_step_mid(double a, double b, void *p, double (*f)(double, void *), 
   }
 
   double rel = 1.;
-  double N = 1000.;
+  int N = 1000;
   double K = 3.;
-  double h = (b-a)/N;
+  double h = (b-a)/(double)N;
   
   double M = h*(*trafo)(*f,a+h/2.,p);
   double M1 = 0.;
@@ -206,9 +206,9 @@ double adapt_step_trap(double a, double b, void *p, double (*f)(double, void *),
   // relative error e>0
 
   double rel = 1.; 		// initialize relative error 
-  double N = 1000.;   		// initialize number of steps to start with for initial stepwidth
+  int N = 1000;   		// initialize number of steps to start with for initial stepwidth
   double K = 2.;   		// initialize halving parameter
-  double h = (b - a) / N;	// initialize stepwidth
+  double h = (b - a) / (double)N;	// initialize stepwidth
   double T = (*f)(a, p) + (*f)(b, p); //analytically evaluated start value 
 
   double T1 = 0.;  // initialize halved stepsize value
@@ -257,8 +257,8 @@ double int_left_riemann(double a, double b, void *p, double (*f)(double, void *)
   }
 
   // double N = 10000.;
-  double N =10000.;
-  double h = (b - a) / N;
+  int N =10000;
+  double h = (b - a) / (double)N;
 
   // left Riemann sum
   double L = (*f)(a, p);           // this bitch is the reason you gotta start with a+h in the for loop
@@ -286,8 +286,8 @@ double int_right_riemann(double a, double b, void *p, double (*f)(double, void *
       return -1. * int_right_riemann(b, a, p, f);
   }
 
-  double N =10000.;
-  double h = (b - a) / N;
+  int N =10000;
+  double h = (b - a) / (double)N;
   double R = 0;
   
   for (int i = 1; i <= N; i ++)
@@ -315,8 +315,8 @@ double int_trapezoidal_double(double a, double b, void *p, double (*f)(double, v
       return -1. * int_trapezoidal_double(b, a, p, f);
   }
 
-  double N =10000.;
-  double h = (b - a) / N;
+  int N =10000;
+  double h = (b - a) / (double)N;
   double T = (*f)(a, p) + (*f)(b, p); //analytically evaluated
 
   
@@ -345,8 +345,8 @@ double int_trapezoidal_int(double a, double b, void *p, double (*f)(double, void
   {
     return -1. * int_trapezoidal_int(b, a, p, f);
   }
-  double N =10000.;
-  double h = (b - a) / N;
+  int N =10000;
+  double h = (b - a) / (double)N;
   double T = (*f)(a, p) + (*f)(b, p); //analytically evaluated
 
   
@@ -375,8 +375,8 @@ double int_simpson_one_loop(double a, double b, void *p, double (*f)(double, voi
       return -1. * int_simpson_one_loop(b, a, p, f);
   }
 
-  double N =100000.;
-  double h = (b - a) / N;
+  int N =100000;
+  double h = (b - a) / (double)N;
   double S = (*f)(a,p)+(*f)(b,p);
   for (int i = 1; i <= N-1; i ++){
 
@@ -409,8 +409,8 @@ double int_simpson_two_loop(double a, double b, void *p, double (*f)(double, voi
       return -1.* int_simpson_two_loop(b, a, p, f);
   }
 
-  double N =100000.;
-  double h = (b - a) / N;
+  int N =100000;
+  double h = (b - a) / (double)N;
   double S = (*f)(a,p)+(*f)(b,p);
   for (int i = 1; i <= N-1; i ++) {
     
@@ -444,8 +444,8 @@ double montecarlo(double a, double b, void *p,  double (*f)(double, void *), dou
   {
       return -1. * montecarlo(b, a, p, f, abe);
   }
-  double N = 1000.;
-  double h = (b-a)/N;
+  int N = 1000;
+  double h = (b-a)/ (double)N;
 
   double result = 0.;
   double error = abe;
@@ -548,9 +548,9 @@ double adapt_step_simp(double a, double b, void *p, double(*f)(double, void *), 
   }
 
   double S = (*f)(a,p)+(*f)(b,p); // intital value, derived analytically
-  double N = 100000.;
+  int N = 100000;
   double K = 2.;
-  double h = (b - a) / N;
+  double h = (b - a) / (double)N;
   double S1 = 0.;
 
   double dummy_subs = 0.; // we need a substitute variable since we derived analytically that the following
